requirements.txt
# FastAPI核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库相关
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# 认证和安全
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# 任务队列
celery==5.3.4
redis==5.0.1

# HTTP客户端和网络工具
httpx==0.25.2
aiohttp==3.9.1
dnspython==2.4.2

# 浏览器自动化
playwright==1.40.0

# 数据处理
pandas==2.1.4
numpy==1.25.2

# 日志和监控
structlog==23.2.0
prometheus-client==0.19.0

# 图像处理
Pillow==10.1.0

# 工具类
python-dotenv==1.0.0
typing-extensions==4.8.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1